requirements.txt
pydantic
jsonpath_ng
python-jsonlogic